Solve ticket #319
In classic theme, if the user does not display the pass icon, group name
and administrateur flag are not display in the session bar, in top of
page.

Change the sessionbar.inc, in classic theme, wit a copy of what is done
from the footer.ing file in the Material Blue theme.